Refactor(woorim/eslint): 각 모델들마다 폴더로 분류
