+++ app/src/middlewares/auth.js
@@ -1,4 +1,4 @@
-const Auth = require("../models/services/Auth");
+const Auth = require("../models/services/Auth/Auth");
 
 // 로그인 된 유저들만 서비스 이용을 허가하는 미들웨어
 const logined = async (req, res, next) => {
