+++ app/src/models/services/Auth/AuthStorage.js
@@ -1,4 +1,4 @@
-const db = require("../../config/db");
+const db = require("../../../config/db");
 
 class AuthStorage {
   static findOneById(id) {
